Initial commit of 96BoardCECase.scad
Under GPL v2.0 license.  This OpenScad file can generate 3D printer
files for cases that fit the 96Boards CE standard.  It can generate
a rectangular case, a rectangular case with curved edges and a rectangular
case with all edges curved.  It can do that for both the normal sized board
as well as the extended sized board.  There is more work to do but its a
good first cut.