Changes to fix the front connectors that stick past the front of the board
Added a variable (front_connector_tolerance_z_drop normally set to .40) to
control how much below the board edge we remove in the final case.  Some
boards the connectors stick out just a bit and with a tight fitting case it
causes problems.  This should fix it.

Edited the README.md file to reflect that some 3D printers will print in
slightly different sized cases, you can use a variable (CE_spec_tolerance,
normally set to .30) to expand or contract the case a little bit if need be.