Some changes to the scad file and added signed off by requirement
Some clean up to the primary scad file.  Lots more to come.  Have determined
that different printers vary a bunch on X and Y sizing they should not but
do.  So going to have to add a multiplyer to the code base to allow for some
printers going oversize and some going undersize.  Have to add % based
multipliers for each access.