Added scaler vars to 96BoardCECase.scad and a TODO file and updated the README
Now all dimensions of a case can be scaled to allow for board or printer
differences.  Also cleaned up some modules a bit more.

Added a TODO file so if anyone is interested in helping it outlines those
areas that can use some work.

Updated the README.md file, corrected spelling errors, and added more info
about the options in the .scad file.